% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{compat}
\alias{compat}
\title{Provide forwards compatibility by converting results from previous versions of CdmOnboarding to the latest version.
Provide backwards compatibility by converting results from the latest version of CdmOnboarding to previous versions.}
\usage{
compat(r, target_version = package_version("3.0.0"))
}
\description{
Provide forwards compatibility by converting results from previous versions of CdmOnboarding to the latest version.
Provide backwards compatibility by converting results from the latest version of CdmOnboarding to previous versions.
}
