% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CdmOnboarding.R
\name{.execute}
\alias{.execute}
\title{Update the HADES package list with:
packageListUrl <- "https://raw.githubusercontent.com/OHDSI/Hades/main/extras/packages.csv"
hadesPackageList <- read.table(packageListUrl, sep = ",", header = TRUE)
packages <- hadesPackageList$name
dump("packages", "")}
\usage{
.execute(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  scratchDatabaseSchema,
  vocabDatabaseSchema,
  oracleTempSchema,
  databaseId,
  databaseName,
  databaseDescription,
  runVocabularyChecks,
  runDataTablesChecks,
  runPerformanceChecks,
  runWebAPIChecks,
  runDedChecks,
  smallCellCount,
  baseUrl,
  sqlOnly,
  outputFolder,
  verboseMode,
  dqdJsonPath,
  optimize,
  dedIngredientIds
)
}
\description{
Update the HADES package list with:
packageListUrl <- "https://raw.githubusercontent.com/OHDSI/Hades/main/extras/packages.csv"
hadesPackageList <- read.table(packageListUrl, sep = ",", header = TRUE)
packages <- hadesPackageList$name
dump("packages", "")
}
