% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDedResults.R
\name{.exportDedResults}
\alias{.exportDedResults}
\title{Export DrugExposureDiagnostics results to csv file}
\usage{
.exportDedResults(results, outputFolder = getwd(), ded_version = "1.0.5")
}
\arguments{
\item{results}{results object from \code{cdmOnboarding}}

\item{outputFolder}{folder to store the results}
}
\description{
Export DrugExposureDiagnostics results to csv file
}
