% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCdmSource.R
\name{.getCdmSource}
\alias{.getCdmSource}
\title{Get CDM source table}
\usage{
.getCdmSource(connectionDetails, cdmDatabaseSchema, outputFolder)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{outputFolder}{Path to store logs and SQL files}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                            A data frame with the CDM source table
}\if{html}{\out{</div>}}
}
\description{
Get CDM source table
}
