% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DqdChecks.R
\name{.processDqdResults}
\alias{.processDqdResults}
\title{Process dqd file}
\usage{
.processDqdResults(dqdJsonPath)
}
\arguments{
\item{dqdJsonPath}{Path to the DQD JSON file}
}
\value{
list with version, overview, startTimestamp, executionTime
}
\description{
Process dqd file
}
