% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeQuery.R
\name{executeQuery}
\alias{executeQuery}
\title{Execute sql file on the given connectionDetails}
\usage{
executeQuery(
  outputFolder,
  sqlFileName,
  successMessage = NULL,
  connectionDetails = NULL,
  sqlOnly = FALSE,
  activeConnection = NULL,
  useExecuteSql = FALSE,
  ...
)
}
\arguments{
\item{outputFolder}{Path to store logs and SQL files}

\item{sqlFileName}{Name of the SQL file to execute}

\item{successMessage}{Message to log when the query is successful}

\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{sqlOnly}{Boolean indicating if only the SQL should be written to file}

\item{activeConnection}{An active connection object to use}

\item{useExecuteSql}{Boolean indicating if the query should be executed using \code{executeSql} instead of \code{querySql}}

\item{...}{Additional parameters to pass to SqlRender::loadRenderTranslateSql}
}
\value{
result of the query
}
\description{
Execute sql file on the given connectionDetails
}
