% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDedResults.R
\name{exportDedResults}
\alias{exportDedResults}
\title{Export DrugExposureDiagnostics results to csv file}
\usage{
exportDedResults(path, outputFolder = getwd())
}
\arguments{
\item{path}{path to the CdmOnboarding .rds results file}

\item{outputFolder}{folder to store the results}
}
\description{
Export DrugExposureDiagnostics results to csv file
}
