% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohortBenchmarkSection.R
\name{generateCohortBenchmarkSection}
\alias{generateCohortBenchmarkSection}
\title{Generates the Cohort Benchmark section for the Results Document}
\usage{
generateCohortBenchmarkSection(doc, df)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{df}{Results object from \code{cdmOnboarding}}
}
\description{
Generates the Cohort Benchmark section for the Results Document
}
