% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataTablesSection.R
\name{generateDataTablesSection}
\alias{generateDataTablesSection}
\title{Generates the Data Tables section for the Results Document}
\usage{
generateDataTablesSection(doc, df, cdmSource, optimized)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{df}{Results object from \code{cdmOnboarding} dataTablesResults}

\item{cdmSource}{CDM source object}

\item{optimized}{boolean indicating if the optimized queries were used}
}
\description{
Generates the Data Tables section for the Results Document
}
