% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDedSection.R
\name{generateDedSection}
\alias{generateDedSection}
\title{Generates the Drug Exposure Diagnostics section for the Results Document}
\usage{
generateDedSection(doc, df)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{df}{Results object from \code{cdmOnboarding}}
}
\description{
Generates the Drug Exposure Diagnostics section for the Results Document
}
