% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDqdSection.R
\name{generateDqdSection}
\alias{generateDqdSection}
\title{Generates the DQD section for the Results Document}
\usage{
generateDqdSection(doc, df)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{df}{Results object from \code{cdmOnboarding}}
}
\description{
Generates the DQD section for the Results Document
}
