% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateExecutionDetailsSection.R
\name{generateExecutionDetails}
\alias{generateExecutionDetails}
\title{Generates the Execution Details section for the Results Document}
\usage{
generateExecutionDetails(doc, df)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{df}{Results object from \code{cdmOnboarding}}
}
\description{
Generates the Execution Details section for the Results Document
}
