% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePerformanceSection.R
\name{generatePerformanceSection}
\alias{generatePerformanceSection}
\title{Generates the Performance section in the Results Document}
\usage{
generatePerformanceSection(doc, results)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{results}{Results object from \code{cdmOnboarding}}
}
\description{
Generates the Performance section in the Results Document
}
