% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVocabularySection.R
\name{generateVocabularySection}
\alias{generateVocabularySection}
\title{Generates the Vocabulary Section in the Results Document}
\usage{
generateVocabularySection(doc, df, smallCellCount)
}
\arguments{
\item{doc}{officer document object to add the section to}

\item{df}{Results object from \code{cdmOnboarding}}

\item{smallCellCount}{number of cells to display in the unmapped/mapped tables}
}
\description{
Generates the Vocabulary Section in the Results Document
}
