% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceChecks.R
\name{getExpectedIndexes}
\alias{getExpectedIndexes}
\title{Default indexes for an OMOP CDM. Note that currently these are the same for both v5.3 and v5.4.}
\usage{
getExpectedIndexes(cdmVersion)
}
\description{
Default indexes for an OMOP CDM. Note that currently these are the same for both v5.3 and v5.4.
}
