% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceChecks.R
\name{getHADESpackages}
\alias{getHADESpackages}
\title{Hard coded list of HADES packages that CdmOnboarding checks against.
Does NOT update automatically when new HADES packages are released.}
\usage{
getHADESpackages()
}
\value{
character vector with HADES package names
}
\description{
Hard coded list of HADES packages that CdmOnboarding checks against.
Does NOT update automatically when new HADES packages are released.
}
