% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DedChecks.R
\name{getMappingLevel}
\alias{getMappingLevel}
\title{Get drug class levels from DrugExposureDiagnostics results}
\usage{
getMappingLevel(dedResults)
}
\arguments{
\item{dedResults}{DrugExposureDiagnostics results object}
}
\value{
data frame with for each ingredient and concept_class_id the number of concepts and records
}
\description{
Get drug class levels from DrugExposureDiagnostics results
}
