% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceChecks.R
\name{performanceChecks}
\alias{performanceChecks}
\title{The performane checks (for v5.x)}
\usage{
performanceChecks(
  connectionDetails,
  resultsDatabaseSchema,
  vocabDatabaseSchema = cdmDatabaseSchema,
  sqlOnly = FALSE,
  outputFolder = "output"
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary.
On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{sqlOnly}{Boolean to determine if Achilles should be fully executed. TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}

\item{outputFolder}{Path to store logs and SQL files}
}
\value{
\preformatted{                            An object of type \code{achillesResults} containing details for connecting to the database containing the results
}
}
\description{
\code{PerformanceChecks} runs a list of performance checks as part of the CDM Onboarding procedure
}
\details{
\code{PerformanceChecks} runs a list of performance checks as part of the CDM Onboarding procedure
}
