% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VocabularyChecks.R
\name{vocabularyChecks}
\alias{vocabularyChecks}
\title{The vocabulary checks (for v5.x)}
\usage{
vocabularyChecks(
  connectionDetails,
  cdmDatabaseSchema,
  vocabDatabaseSchema = cdmDatabaseSchema,
  smallCellCount = 5,
  sqlOnly = FALSE,
  outputFolder = "output",
  optimize = FALSE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{vocabDatabaseSchema}{String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema.
On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.}

\item{smallCellCount}{To avoid patient identifiability, cells with small counts (<= smallCellCount) are deleted. Set to NULL if you don't want any deletions.}

\item{sqlOnly}{Boolean to determine if Achilles should be fully executed. TRUE = just generate SQL files, don't actually run, FALSE = run Achilles}

\item{outputFolder}{Path to store logs and SQL files}

\item{optimize}{Boolean to determine if heuristics will be used to speed up execution. Currently only implemented for postgresql databases. Default = FALSE}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                            An object of type \code{achillesResults} containing details for connecting to the database containing the results
}\if{html}{\out{</div>}}
}
\description{
\code{vocabularyChecks} runs a list of checks on the vocabulary as part of the CDM Onboarding procedure
}
\details{
\code{vocabularyChecks} runs a list of checks on the vocabulary as part of the CDM Onboarding procedure
}
