% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{compat}
\alias{compat}
\title{Compatibility of CdmOnboarding results}
\usage{
compat(r, target_version = package_version("3.0"))
}
\arguments{
\item{r}{A list of results from CdmOnboarding.}

\item{target_version}{The version to convert the results to, only v3 supported.}
}
\value{
A list of results from CdmOnboarding, forwards compatible with v3.
}
\description{
\code{compat} provides compatibility by converting results from previous versions of CdmOnboarding to the latest version.
}
