% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCdmConnection.R
\name{.getCdmConnection}
\alias{.getCdmConnection}
\title{Convenience function to get a CDMConnector connection from connectionDetails}
\usage{
.getCdmConnection(connectionDetails)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}
}
\value{
CDMConnector connection
}
\description{
Convenience function to get a CDMConnector connection from connectionDetails
}
