% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDedResults.R
\name{exportDedResults}
\alias{exportDedResults}
\title{Export DrugExposureDiagnostics results to csv file in same folder as input file path}
\usage{
exportDedResults(path)
}
\arguments{
\item{path}{path to the CdmOnboarding .rds results file, output is written to the same folder}
}
\description{
Export DrugExposureDiagnostics results to csv file in same folder as input file path
}
