% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateResultsDocument.R
\name{generateResultsDocument}
\alias{generateResultsDocument}
\title{Generates the Results Document}
\usage{
generateResultsDocument(results, outputFolder, authors)
}
\arguments{
\item{results}{Results object from \code{cdmOnboarding}}

\item{outputFolder}{Folder to store the results}

\item{authors}{List of author names to be added in the document}
}
\description{
\code{generateResultsDocument} creates a word document with results based on a template
}
